requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.3.0
arxiv==1.4.7
asgiref==3.6.0
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
dataclasses-json==0.5.7
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.6
distro==1.9.0
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-import-export==3.3.7
django-simpleui==2023.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx2txt==0.8
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
feedparser==6.0.10
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
isort==5.12.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.0.159
lazy-object-proxy==1.9.0
MarkupPy==1.14
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
platformdirs==3.5.0
psycopg==3.1.13
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pylint==2.17.3
PyMuPDF==1.22.2
pypdf==3.8.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
referencing==0.33.0
regex==2023.5.5
requests==2.27.1
requests-oauthlib==1.3.1
rpds-py==0.18.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.10.0
urllib3==1.26.15
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
